Only disallow mining with sword
It turns out the checks provided by forge for 'can harvest block?' are
not consistent (or maybe I don't understand what they're doing :)).
And I'm pretty sure it's only swords that can't mine things in creative
mode. I guess the implication is that in survival, if you try to mine
something with an inadequate tool, you'll have to figure it out yourself
by observing that nothing breaks...